+++ proto/src/prost/ibc.mock.rs
@@ -1,15 +1,17 @@
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Header {
     #[prost(message, optional, tag="1")]
-    pub height: ::std::option::Option<super::core::client::v1::Height>,
+    pub height: ::core::option::Option<super::core::client::v1::Height>,
+    #[prost(uint64, tag="2")]
+    pub timestamp: u64,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ClientState {
     #[prost(message, optional, tag="1")]
-    pub header: ::std::option::Option<Header>,
+    pub header: ::core::option::Option<Header>,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ConsensusState {
     #[prost(message, optional, tag="1")]
-    pub header: ::std::option::Option<Header>,
+    pub header: ::core::option::Option<Header>,
 }